pyproject.toml
[project]
name = "backend-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.133.1",
    "google-genai>=1.65.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.41.0",
]
